emotion aware tutor/backend/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.10.5
click==8.3.0
colorama==0.4.6
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.111.0
fastapi-cli==0.0.13
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
orjson==3.11.3
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
Pygments==2.19.2
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.3
rich==14.2.0
rich-toolkit==0.15.1
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.34
starlette==0.37.2
typer==0.19.2
typing_extensions==4.15.0
ujson==5.11.0
uvicorn==0.30.1
watchfiles==1.1.0
websockets==15.0.1

# SER dependencies
numpy<2.0  # opensmile requires NumPy 1.x compatibility
scipy
soundfile
librosa
scikit-learn
joblib
opensmile

# FER dependencies
tensorflow>=2.16.0  # Use 2.16+ for protobuf 4.x compatibility
protobuf>=3.20.0  # TensorFlow 2.16+ supports protobuf 4.x
Pillow